chewing through some of the final test fixtures
some very gnarly html and json fixtures are providing a few final edge
cases to solve; there may be more but I'm finding the line where I think
I don't want to support certain edge cases or grey-area structures.
6 erroring tests in the nasty json fixture; 5 of them are acceptable
errors that are for unsupported syntax.
the 6th error is for "" as an array item; it serializes to nothing in
htmnl and then parses back as an empty _ii node, causing an error. This
needs to be solved but I've got three different thoughts on how to solve
it:
1 :  _VOID clusters with _content: never (requiring some kind of retyping
of HsonNode to accommodate the new data type, or an exceptional type
somehow)
2 : _ii nodes can have 0 children but must carry a
`_meta.data-_empty="true"`` property (I don't love data-_empty because
you'd ahve to remove it when you add content and so check for it every
time--possibly)
3:  _ii nodes can have 0 children; that's it. no _meta changes. only
issue here is the risk of silent data loss--feels small but probably
unacceptable.

there's time to consider; the next and final step in the
refactor will be liveTree and then it's clean-up and resume
work on the test showcase